Refactored package DataAccess and DAO
Now called data_access and dao respectively
