+++ app/src/main/java/com/bimapp/ProjectsViewActivity.java
@@ -33,16 +33,16 @@ import com.bimapp.controller.FragmentTopic;
 import com.bimapp.controller.FragmentTopicList;
 import com.bimapp.controller.interfaces.ProjectsFragmentInterface;
 import com.bimapp.model.ImageFile;
+import com.bimapp.model.data_access.entityManagers.IssueBoardExtensionsEntityManager;
+import com.bimapp.model.data_access.entityManagers.ProjectEntityManager;
+import com.bimapp.model.data_access.network.APICall;
+import com.bimapp.model.data_access.network.Callback;
+import com.bimapp.model.data_access.network.NetworkConnManager;
 import com.bimapp.model.entity.IssueBoardExtensions;
 import com.bimapp.model.entity.Project;
 import com.bimapp.model.entity.Template.Template;
 import com.bimapp.model.entity.Topic;
 import com.bimapp.model.entity.User;
-import com.bimapp.model.DataAccess.entityManagers.IssueBoardExtensionsEntityManager;
-import com.bimapp.model.DataAccess.entityManagers.ProjectEntityManager;
-import com.bimapp.model.DataAccess.network.APICall;
-import com.bimapp.model.DataAccess.network.Callback;
-import com.bimapp.model.DataAccess.network.NetworkConnManager;
 
 import org.json.JSONException;
 import org.json.JSONObject;
