+++ app/src/main/java/com/bimapp/controller/FragmentNewComment.java
@@ -12,9 +12,9 @@ import android.view.ViewGroup;
 
 import com.bimapp.BimApp;
 import com.bimapp.controller.interfaces.CommentFragmentInterface;
+import com.bimapp.model.data_access.entityManagers.CommentEntityManager;
 import com.bimapp.model.entity.Comment;
 import com.bimapp.model.entity.Topic;
-import com.bimapp.model.DataAccess.entityManagers.CommentEntityManager;
 import com.bimapp.view.CommentView;
 import com.bimapp.view.interfaces.CommentViewInterface;
