+++ app/src/main/java/com/bimapp/controller/interfaces/NewTopicFragmentInterface.java
@@ -5,19 +5,19 @@ import com.bimapp.model.entity.Topic;
 /**
  * Interface to communicate between the {@link com.bimapp.controller.FragmentNewTopic} responsible
  * for allowing users to post new topics, and the
- * {@link com.bimapp.model.DataAccess.entityManagers.TopicsEntityManager} which is responsible for
+ * {@link com.bimapp.model.data_access.entityManagers.TopicsEntityManager} which is responsible for
  * posting the topic to the API.
  *
  * {@link com.bimapp.controller.FragmentNewTopic} must implement this interface
  */
 public interface NewTopicFragmentInterface {
     /**
-     * This interface is implemented by the {@link com.bimapp.model.DataAccess.entityManagers.TopicsEntityManager}
+     * This interface is implemented by the {@link com.bimapp.model.data_access.entityManagers.TopicsEntityManager}
      * and
      */
     interface NewTopicFragmentListener{
         /**
-         * Method telling the {@link com.bimapp.model.DataAccess.entityManagers.TopicsEntityManager} to post a
+         * Method telling the {@link com.bimapp.model.data_access.entityManagers.TopicsEntityManager} to post a
          * topic to the api. Called from {@link com.bimapp.controller.FragmentNewTopic}
          * @param topic The topic you want to post
          */
@@ -26,7 +26,7 @@ public interface NewTopicFragmentInterface {
 
     /**
      * Method implemented by {@link com.bimapp.controller.FragmentNewTopic} and is called when
-     * {@link com.bimapp.model.DataAccess.entityManagers.TopicsEntityManager} has finished posting a topic.
+     * {@link com.bimapp.model.data_access.entityManagers.TopicsEntityManager} has finished posting a topic.
      */
     void postedTopic(boolean success, Topic topic);
 }
