+++ app/src/main/java/com/bimapp/model/data_access/DataProvider.java
@@ -1,4 +1,4 @@
-package com.bimapp.model.DataAccess;
+package com.bimapp.model.data_access;
 
 import android.content.ContentProvider;
 import android.content.ContentValues;
@@ -6,8 +6,6 @@ import android.database.Cursor;
 import android.net.Uri;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-
-import com.bimapp.BimApp;
 
 public class DataProvider extends ContentProvider {
     public AppDatabase database;
