+++ app/src/main/java/com/bimapp/model/data_access/dao/UserDAO.java
@@ -1,7 +1,6 @@
-package com.bimapp.model.DataAccess.DAO;
+package com.bimapp.model.data_access.dao;
 
 import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Delete;
 import android.arch.persistence.room.Insert;
 import android.arch.persistence.room.OnConflictStrategy;
 import android.arch.persistence.room.Query;
