+++ app/src/main/java/com/bimapp/model/data_access/entityManagers/ProjectEntityManager.java
@@ -1,15 +1,15 @@
-package com.bimapp.model.DataAccess.entityManagers;
+package com.bimapp.model.data_access.entityManagers;
 
 import android.util.Log;
 
 import com.android.volley.Request;
 import com.bimapp.BimApp;
 import com.bimapp.controller.interfaces.ProjectsFragmentInterface;
+import com.bimapp.model.data_access.network.APICall;
+import com.bimapp.model.data_access.network.Callback;
+import com.bimapp.model.data_access.network.NetworkConnManager;
 import com.bimapp.model.entity.EntityListConstructor;
 import com.bimapp.model.entity.Project;
-import com.bimapp.model.DataAccess.network.APICall;
-import com.bimapp.model.DataAccess.network.Callback;
-import com.bimapp.model.DataAccess.network.NetworkConnManager;
 
 import org.json.JSONArray;
 import org.json.JSONException;
