+++ app/src/main/java/com/bimapp/model/data_access/entityManagers/TopicsEntityManager.java
@@ -1,4 +1,4 @@
-package com.bimapp.model.DataAccess.entityManagers;
+package com.bimapp.model.data_access.entityManagers;
 
 import android.util.Log;
 
@@ -7,11 +7,11 @@ import com.bimapp.BimApp;
 import com.bimapp.controller.interfaces.NewTopicFragmentInterface;
 import com.bimapp.controller.interfaces.TopicFragmentInterface;
 import com.bimapp.controller.interfaces.TopicsFragmentInterface;
+import com.bimapp.model.data_access.network.APICall;
+import com.bimapp.model.data_access.network.Callback;
+import com.bimapp.model.data_access.network.NetworkConnManager;
 import com.bimapp.model.entity.EntityListConstructor;
 import com.bimapp.model.entity.Topic;
-import com.bimapp.model.DataAccess.network.APICall;
-import com.bimapp.model.DataAccess.network.Callback;
-import com.bimapp.model.DataAccess.network.NetworkConnManager;
 
 import org.json.JSONArray;
 import org.json.JSONException;
