+++ app/src/main/java/com/bimapp/model/data_access/network/BimAppRequest.java
@@ -1,4 +1,4 @@
-package com.bimapp.model.DataAccess.network;
+package com.bimapp.model.data_access.network;
 
 import android.graphics.Bitmap;
 import android.support.annotation.Nullable;
