+++ app/src/main/java/com/bimapp/model/data_access/network/Callback.java
@@ -1,4 +1,4 @@
-package com.bimapp.model.DataAccess.network;
+package com.bimapp.model.data_access.network;
 
 /**
  * Created by zorri on 14/03/2018.
