+++ app/src/main/java/com/bimapp/model/data_access/network/oauth/OAuthCallback.java
@@ -1,7 +1,7 @@
-package com.bimapp.model.DataAccess.network.oauth;
+package com.bimapp.model.data_access.network.oauth;
 
 import com.android.volley.VolleyError;
-import com.bimapp.model.DataAccess.network.Callback;
+import com.bimapp.model.data_access.network.Callback;
 
 /**
  * Created by Håkon on 13.03.2018.
